+++ view-inspector-plugin/src/main/groovy/view_inspector/plugin/internal/LayoutResourceParser.groovy
@@ -22,12 +22,17 @@ import com.android.annotations.NonNull
 class LayoutResourceParser {
   private LayoutResourceParser() {}
 
-  public static Set<String> parse(@NonNull File layoutFile) {
+  public static Set<String> parse(@NonNull File layoutFile, String[] excludePackages) {
     Set<String> viewClassNames = new HashSet<String>()
 
     def rootNode = new XmlParser().parse(layoutFile)
     traverseLayoutXml(rootNode, viewClassNames)
 
+    excludePackages.each { excludePackage ->
+      viewClassNames.removeAll {
+        it.startsWith(excludePackage)
+      }
+    }
     return viewClassNames
   }
 
@@ -54,6 +59,8 @@ class LayoutResourceParser {
     def className
     if (name == 'View') {
       className = "android.view.View"
+    } else if (name == 'SurfaceView') {
+      className = "android.view.SurfaceView"
     } else if (name == 'WebView') {
       className = "android.webkit.WebView"
     } else if (
@@ -62,7 +69,9 @@ class LayoutResourceParser {
         name == "ViewStub" ||
         name == "fragment" ||
         name == "requestLayout" ||
-        name == "DateTimeView") {
+        name == "DateTimeView" ||
+        name == "Space" ||
+        name == "requestFocus") {
       className = null
     } else if (name.indexOf('.') == -1) {
       className = "android.widget.${name}"