+++ view-inspector-runtime/src/main/java/view_inspector/dagger/ActivityComponent.java
@@ -1,24 +1,27 @@
 package view_inspector.dagger;
 
 import dagger.Component;
+import view_inspector.ViewInspector;
 import view_inspector.dagger.scope.PerActivity;
 import view_inspector.probe.ViewInspectorInterceptor;
 import view_inspector.ui.ViewInspectorToolbar;
 import view_inspector.ui.dialog.BaseDialog;
 import view_inspector.ui.dialog.ProfileResultDialog;
 import view_inspector.ui.dialog.ProfileSettingDialog;
+import view_inspector.ui.dialog.SetViewFilterDialog;
 import view_inspector.ui.dialog.adapter.ProfileResultAdapter;
+import view_inspector.ui.dialog.adapter.ViewFilterAdapter;
 import view_inspector.ui.dialog.adapter.ViewRootAdapter;
 import view_inspector.ui.menu.BaseMenu;
 import view_inspector.ui.menu.BoundaryMenu;
 import view_inspector.ui.menu.EventMenu;
 import view_inspector.ui.menu.LayerMenu;
 import view_inspector.ui.menu.SettingsMenu;
-import view_inspector.weaving.ViewInspectorAspect;
 
 @PerActivity @Component(dependencies = ApplicationComponent.class, modules = ActivityModule.class)
 public interface ActivityComponent {
-  void inject(ViewInspectorAspect viewInspectorAspect);
+
+  void inject(ViewInspector viewInspector);
 
   void inject(ViewInspectorInterceptor viewInspectorInterceptor);
 
@@ -43,4 +46,8 @@ public interface ActivityComponent {
   void inject(ProfileResultDialog profileResultDialog);
 
   void inject(ProfileSettingDialog profileSettingDialog);
+
+  void inject(ViewFilterAdapter viewFilterAdapter);
+
+  void inject(SetViewFilterDialog setViewFilterDialog);
 }