+++ CHANGELOG.md
@@ -1,10 +1,46 @@
 Change Log
 ==========
 
+Version 0.2.0 *(2015-11-17)*
+----------------------------
+
+ * Aspect related code are now separate modules
+ * Plugin is able to pick modules according to application's API level
+ * Marshmallow is now supported
+ * Fix: Resource files conflict (#49)
+ * Fix: Handle ActionBarOverlayLayout as viewParent case (#43)
+ * Fix: Skip requestFocus (#41)
+ * Fix: Add android.internals to isNotSupportedViewClass (#35)
+ * Fix: Handle resource name not found exception (#34)
+
+
+Version 0.1.4 *(2015-10-29)*
+----------------------------
+
+ * Fix: Crashes on Pre-Lollipop due to Vector Drawables
+ * Fix: Correct minSdkVersion to 14
+
+
+Version 0.1.3 *(2015-10-29)*
+----------------------------
+
+ * Support exclude packages option
+ * View filter can now be set in Settings Menu
+ * Fix: Several UI bugs
+
+
+Version 0.1.2 *(2015-10-28)*
+----------------------------
+
+ * Fix: Issues with hard-coded Manifest file path
+ * Fix: Wrong class name resolved with WebView
+ * Fix: Several UI bugs
+
+
 Version 0.1.1 *(2015-10-26)*
 ----------------------------
 
- * Works with compileSdkVersion >= 21
+ * Works with compileSdkVersion >= 21 but not Marshmallow
  * Fix: Attempting to assign weaker access privileges
  * Fix: Several UI bugs